Fix more test cases
