+++ distributed/core.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
 import asyncio
+import contextvars
 import inspect
 import logging
 import sys
 import threading
 import traceback
 import uuid
-import warnings
 import weakref
 from collections import defaultdict
 from collections.abc import Container
@@ -37,8 +37,10 @@ from distributed.comm import (
 from distributed.metrics import time
 from distributed.system_monitor import SystemMonitor
 from distributed.utils import (
+    Handler,
     TimeoutError,
     get_traceback,
+    handler_factory,
     has_keyword,
     is_coroutine_function,
     recursive_to_dict,
@@ -96,75 +98,6 @@ tick_maximum_delay = parse_timedelta(
 )
 
 LOG_PDB = dask.config.get("distributed.admin.pdb-on-err")
-
-
-class Handler:
-    def __init__(self, func):
-        if isinstance(func, Handler):
-            func = func.func
-        elif callable(func):
-            pass
-        else:
-            raise TypeError(f"{func} must be callable")
-
-        assert callable(func)
-
-        self.func = func
-        self.stimulus_name = func.__name__.replace("_", "-")
-        self.sig = inspect.signature(func)
-        self.wants_stimulus_id = "stimulus_id" in self.sig.parameters
-
-    def __call__(self, *args, **kwargs):
-        if self.wants_stimulus_id and "stimulus_id" not in kwargs:
-            kwargs["stimulus_id"] = f"{self.stimulus_name}-{time()}"
-
-        return self.func(*args, **kwargs)
-
-    def __hash__(self):
-        return hash(self.func)
-
-    def expects_comm(self):
-        params = list(self.sig.parameters.keys())
-
-        if params and params[0] == "comm":
-            return True
-        if params and params[0] == "stream":
-            warnings.warn(
-                "Calling the first arugment of a RPC handler `stream` is "
-                "deprecated. Defining this argument is optional. Either remove the "
-                f"arugment or rename it to `comm` in {self.func}{self.sig}.",
-                FutureWarning,
-            )
-            return True
-        return False
-
-    def __reduce__(self):
-        return (handler_factory, (self.func,))
-
-    def __name__(self):
-        return self.func.__name__
-
-    def __eq__(self, other: object):
-        return isinstance(other, Handler) and self.func == other.func
-
-    def __str__(self):
-        return f"handler({self.func})"
-
-    def __repr__(self):
-        return f"handler({self.func})"
-
-
-class AsyncHandler(Handler):
-    async def __call__(self, *args, **kwargs):
-        return await super().__call__(*args, **kwargs)
-
-    # def __eq__(self, other: AsyncHandler):
-    #     return isinstance(other, AsyncHandler) and self.func == other.func
-
-
-def handler_factory(func):
-    cls = AsyncHandler if is_coroutine_function(func) else Handler
-    return cls(func)
 
 
 class Server:
@@ -240,6 +173,7 @@ class Server:
             k: handler_factory(f) for k, f in self.stream_handlers.items()
         }
 
+        self.context = contextvars.copy_context()
         self.id = type(self).__name__ + "-" + str(uuid.uuid4())
         self._address = None
         self._listen_address = None
